requirements.txt
appdirs==1.4.2
CoffeeScript==1.1.1
Django==1.8.3
django-appconf==1.0.2
django-braces==1.4.0
django-cors-headers==1.1.0
django-debug-toolbar==1.3.0
django-extensions==1.4.5
django-model-utils==2.2
django-pipeline==1.3.26
django-static-precompiler==0.9
django-suit==0.2.13
django-tinymce==2.0.4
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
docutils==0.12
drf-generators==0.2.4
futures==3.0.1
gevent==1.0.1
gnureadline==6.3.3
greenlet==0.4.6
html5lib==0.9999999
httpretty==0.8.14
ipython==2.3.0
iso3166==0.8
Jinja2==2.7.3
jsonfield==1.0.3
logutils==0.3.3
lxml==3.6.4
MarkupSafe==0.23
packaging==16.8
pdfcrowd==2.7
pdfkit==0.5.0
Pillow==2.8.1
pinax-stripe==3.3.0
pisa==3.0.33
psycopg2==2.7
PyExecJS==1.1.0
Pygments==2.0.2
PyJWT==1.4.0
pyparsing==2.2.0
PyPDF2==1.26.0
pytz==2016.7
reportlab==3.2.0
requests==2.11.1
requirements==0.1
schwifty==2.0.1
signaturit-sdk==1.0.0
six==1.9.0
Sphinx==1.2.3
sphinx-rtd-theme==0.1.6
sqlparse==0.1.15
stripe==1.41.1
# uWSGI==2.0.4
xhtml2pdf==0.0.6
